update : 전 commit에 수행하지 않은 ClubMemberService클래스 분리 및 execute 컨벤션 적용
