+++ src/main/kotlin/team/incube/flooding/domain/club/service/ExileMemberService.kt
@@ -1,0 +1,8 @@
+package team.incube.flooding.domain.club.service
+
+interface ExileMemberService {
+    fun execute(
+        clubId: Long,
+        userId: Long,
+    )
+}
