+++ src/main/kotlin/team/incube/flooding/domain/club/service/InviteMemberService.kt
@@ -1,0 +1,8 @@
+package team.incube.flooding.domain.club.service
+
+interface InviteMemberService {
+    fun execute(
+        clubId: Long,
+        userId: Long,
+    )
+}
