+++ src/main/kotlin/team/incube/flooding/domain/club/service/TransferOwnershipService.kt
@@ -1,0 +1,8 @@
+package team.incube.flooding.domain.club.service
+
+interface TransferOwnershipService {
+    fun execute(
+        clubId: Long,
+        targetUserId: Long,
+    )
+}
